Separate sql probes
