+++ bpf/go_sql.h
@@ -1,0 +1,114 @@
+// Copyright The OpenTelemetry Authors
+// Copyright Grafana Labs
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// This implementation was inspired by https://github.com/open-telemetry/opentelemetry-go-instrumentation/blob/ca1afccea6ec520d18238c3865024a9f5b9c17fe/internal/pkg/instrumentors/bpf/database/sql/bpf/probe.bpf.c
+// and has been modified since.
+
+typedef struct sql_func_invocation {
+    u64 start_monotime_ns;
+    u64 sql_param;
+    u64 query_len;
+    tp_info_t tp;
+} sql_func_invocation_t;
+
+struct {
+    __uint(type, BPF_MAP_TYPE_LRU_HASH);
+    __type(key, void *); // key: pointer to the request goroutine
+    __type(value, sql_func_invocation_t);
+    __uint(max_entries, MAX_CONCURRENT_REQUESTS);
+} ongoing_sql_queries SEC(".maps");
+
+static __always_inline void set_sql_info(void *goroutine_addr, void *sql_param, void *query_len) {
+    sql_func_invocation_t invocation = {
+        .start_monotime_ns = bpf_ktime_get_ns(),
+        .sql_param = (u64)sql_param,
+        .query_len = (u64)query_len,
+        .tp = {0}
+    };
+
+    // We don't look up in the headers, no http/grpc request, therefore 0 as last argument
+    client_trace_parent(goroutine_addr, &invocation.tp, 0);
+
+    // Write event
+    if (bpf_map_update_elem(&ongoing_sql_queries, &goroutine_addr, &invocation, BPF_ANY)) {
+        bpf_dbg_printk("can't update map element");
+    }
+}
+
+SEC("uprobe/queryDC")
+int uprobe_queryDC(struct pt_regs *ctx) {
+    bpf_dbg_printk("=== uprobe/queryDC === ");
+    void *goroutine_addr = GOROUTINE_PTR(ctx);
+    bpf_dbg_printk("goroutine_addr %lx", goroutine_addr);
+
+    void *sql_param = GO_PARAM8(ctx);
+    void *query_len = GO_PARAM9(ctx);
+    set_sql_info(goroutine_addr, sql_param, query_len);
+    return 0;
+}
+
+SEC("uprobe/execDC")
+int uprobe_execDC(struct pt_regs *ctx) {
+    bpf_dbg_printk("=== uprobe/execDC === ");
+    void *goroutine_addr = GOROUTINE_PTR(ctx);
+    bpf_dbg_printk("goroutine_addr %lx", goroutine_addr);
+
+    void *sql_param = GO_PARAM6(ctx);
+    void *query_len = GO_PARAM7(ctx);
+    set_sql_info(goroutine_addr, sql_param, query_len);
+    return 0;
+}
+
+SEC("uprobe/queryDC")
+int uprobe_queryReturn(struct pt_regs *ctx) {
+
+    bpf_dbg_printk("=== uprobe/query return === ");
+    void *goroutine_addr = GOROUTINE_PTR(ctx);
+    bpf_dbg_printk("goroutine_addr %lx", goroutine_addr);
+
+    sql_func_invocation_t *invocation = bpf_map_lookup_elem(&ongoing_sql_queries, &goroutine_addr);
+    if (invocation == NULL) {
+        bpf_dbg_printk("Request not found for this goroutine");
+        return 0;
+    }
+    bpf_map_delete_elem(&ongoing_sql_queries, &goroutine_addr);
+
+    sql_request_trace *trace = bpf_ringbuf_reserve(&events, sizeof(sql_request_trace), 0);
+    if (trace) {
+        task_pid(&trace->pid);
+        trace->type = EVENT_SQL_CLIENT;
+        trace->start_monotime_ns = invocation->start_monotime_ns;
+        trace->end_monotime_ns = bpf_ktime_get_ns();
+
+        void *resp_ptr = GO_PARAM1(ctx);
+        trace->status = (resp_ptr == NULL);
+        trace->tp = invocation->tp;
+
+        u64 query_len = invocation->query_len;
+        if (query_len > sizeof(trace->sql)) {
+            query_len = sizeof(trace->sql);
+        }
+        bpf_probe_read(trace->sql, query_len, (void*)invocation->sql_param);
+        bpf_dbg_printk("Found sql statement %s", trace->sql);
+        if (query_len < sizeof(trace->sql)) {
+            trace->sql[query_len] = 0;
+        }
+        // submit the completed trace via ringbuffer
+        bpf_ringbuf_submit(trace, get_flags());
+    } else {
+        bpf_dbg_printk("can't reserve space in the ringbuffer");
+    }
+    return 0;
+}
